requirements-2.txt
ansible==1.9.2
cached-property==1.2.0
Cython==0.22.1
decorator==3.4.2
ecdsa==0.13
filemagic==1.6
FITS-tools
Flask==0.10.1
Flask-JWT==0.2.0
Flask-Migrate==1.4.0
Flask-Script==2.0.5
Flask-SocketIO==0.6.0
Flask-SQLAlchemy==2.0
Flask-uWSGI-WebSocket==0.4.1
gc3pie==2.4
geojson==1.3.0
gnureadline==6.3.3
hashids==1.1.0
h5py==2.5.0
image-registration==0.2.1
ipdb==0.8.1
ipython==3.2.0
julia==0.1.1
lxml==3.4.4
mahotas==1.4.0
matlab-wrapper==0.9.6
matplotlib==1.4.3
mock==1.0.1
mpld3==0.2
natsort==4.0.3
networkx==1.9.1
nose==1.3.7
openslide-python-1.1.0
pandas==0.16.2
paramiko==1.15.2
parsedatetime==1.5
passlib-1.6.4
Pillow==2.9.0
prettytable==0.7.2
psycopg2-2.6.1
pyCLI==2.0.3
pycrypto==2.6.1
pyparsing==2.0.3
pypng==0.0.17
python-dateutil==2.4.2
pytest==2.7.2
pytz==2015.4
PyYAML==3.11
scikit-image==0.11.3
scikit-learn==0.16.0
Shapely==1.5.9
six==1.9.0
sphinxcontrib-matlabdomain==0.2.7
sphinx-pyreverse==0.0.12
SQLAlchemy==1.0.6
tifffile==0.6.2
subprocess32==3.2.6
rpy2==2.6.0
ruamel.yaml==0.8
